ai-model-api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.20.0
opencv-python-headless>=4.5.0
scikit-learn>=1.5.0
